+++ R/benchmark_matrix_functions.R
@@ -24,7 +24,7 @@ bm_matrix_fun_fft = function(runs=3, verbose=TRUE) {
   for (i in 1:runs) {
     a = Rnorm(2500000)
     invisible(gc())
-    timings[i, 1:3] = system.time({b <- fft(a)})[1:3]
+    timings[i, 1:3] = system.time({b <- fft(a)})[1:3] #nolint
   }
   if (verbose)
     message(c("\tFFT over 2,500,000 random values", timings_mean(timings)))
@@ -43,7 +43,7 @@ bm_matrix_fun_eigen = function(runs=3, verbose=TRUE) {
     a = array(Rnorm(600 * 600), dim = c(600, 600))
     invisible(gc())
     timings[i, 1:3] = system.time({
-      b <- eigen(a, symmetric = FALSE, only.values = TRUE)$Value})[1:3]
+      b <- eigen(a, symmetric = FALSE, only.values = TRUE)$Value})[1:3] #nolint
   }
   if (verbose)
     message(c("\tEigenvalues of a 640 x 640 random matrix", timings_mean(timings)))
@@ -60,7 +60,7 @@ bm_matrix_fun_determinant = function(runs = 3, verbose = TRUE) {
   for (i in 1:runs) {
     a = Rnorm(2500 * 2500); dim(a) = c(2500, 2500)
     invisible(gc())
-    timings[i, 1:3] = system.time({b <- det(a)})[1:3]
+    timings[i, 1:3] = system.time({b <- det(a)})[1:3] #nolint
   }
   if (verbose)
     message(c("\tDeterminant of a 2,500 x 2,500 random matrix", timings_mean(timings)))
@@ -79,7 +79,7 @@ bm_matrix_fun_cholesky = function(runs = 3, verbose = TRUE) {
     a = crossprod(new("dgeMatrix", x = Rnorm(3000 * 3000),
                        Dim = as.integer(c(3000, 3000))))
     invisible(gc())
-    timings[i, 1:3] = system.time({b <- chol(a)})[1:3]
+    timings[i, 1:3] = system.time({b <- chol(a)})[1:3] #nolint
   }
   if (verbose)
     message(c("\tCholesky decomposition of a 3,000 x 3,000 matrix", timings_mean(timings)))
@@ -96,7 +96,7 @@ bm_matrix_fun_inverse = function(runs=3, verbose=TRUE) {
   for (i in 1:runs) {
     a = new("dgeMatrix", x = Rnorm(1600 * 1600), Dim = as.integer(c(1600, 1600)))
     invisible(gc())
-    timings[i, 1:3] = system.time({b <- solve(a)})[1:3]
+    timings[i, 1:3] = system.time({b <- solve(a)})[1:3] #nolint
   }
   if (verbose)
     message(c("\tInverse of a 1,600 x 1,600 random matrix", timings_mean(timings)))